requirements.txt
aniso8601==4.0.1
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.0.1
Click==7.0
decorator==4.3.0
dnspython==1.15.0
docx==0.2.4
feedparser==5.2.1
Flask==1.0.2
Flask-CAS==1.0.0
Flask-PyMongo==2.2.0
flask-restplus==0.12.1
idna==2.7
itsdangerous==1.1.0
jaraco.functools==1.20
Jinja2==2.10
jsonschema==2.6.0
lxml==4.2.5
MarkupSafe==1.1.0
more-itertools==4.3.0
-e git+git@github.com:materialsproject/MPContribs.git@075f20a696052eab1a26d80968e6e706fe9545f2#egg=mpcontribs
pattern3==3.0.0
pdfminer.six==20181108
pdfminer3k==1.3.1
Pillow==5.3.0
pluggy==0.8.0
ply==3.11
portend==2.3
py==1.7.0
pycryptodome==3.7.0
pymongo==3.7.2
pytest==4.0.0
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
sortedcontainers==2.0.5
tempora==1.14
urllib3==1.24.1
Werkzeug==0.14.1
xmltodict==0.11.0
zc.lockfile==1.4
